add authorisation callback handler
The auth server is responsible for receiving the callback from Spotify during the authorisation flow. After the authorisation URL is generated for the user, the authorisation server is started which waits for the Spotify API redirection. When the callback is received the auth server extracts the authorisation token which is returned for use in the next authorisation flow stage.